// api-sankhya-center/src/sankhya/tgfgru/models/tgfgru.interface.ts
export interface Tgfgru {
  codgrupoprod: number
  descrgrupoprod: string
  codgrupai: number
  analitico: string
  grau: number
  limcurva_b?: number
  limcurva_c?: number
  comcurva_a?: number
  comcurva_b?: number
  comcurva_c?: number
  particmeta?: number
  metaqtd?: number
  percmetacontrb?: number
  areaocupunid?: number
  qtdexposicao?: number
  valest: string
  imagem?: Buffer
  grupoicms?: number
  ativo: string
  codnat: number
  codcencus: number
  codproj: number
  solcompra: string
  percmeta?: number
  corfundo?: string
  corfonte?: string
  regrawms: string
  pedirlib?: string
  inftrapezio: string
  temflv: string
  aprprodvda: string
  codrfa?: number
  agrupalocvalest: string
  perccmtnac: number
  perccmtimp: number
  aliqnac?: number
  aliqimp?: number
  perccmtfed: number
  perccmtest: number
  perccmtmun: number
  visivelappos?: string
  codctactbefd?: number
  forcaexpeconect?: string
  nuversao?: number
  consgruprodcat42?: string
  dhalter?: Date
  codusu?: number
  calrupturaestoque?: string
}